Heiko Schulz:
Today I finished a first version of an ASK 21 Glider.
It's a very wellknown glider, over 800s are built.

There are still a lot of things to do: some
instruments are missing, the pilots are missing and
there is a some artefacts on the canopy ( in OSG is
much nicer!).

The YASim-FDM is lacking - I don't know how to match
this to the real perfomance...